HavceBeginner/challenge/requirements.txt
Flask==3.1.0
pycryptodome==3.21.0
requests==2.32.3
gunicorn==20.1.0
gevent